backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-cors==3.0.10
python-dotenv==1.0.0
psycopg2-binary==2.9.6